fix peak hold bug in issue #102

+++ src/comp/VuMeters.hpp
@@ -127,7 +127,7 @@ struct VuMeterBase : OpaqueWidget {
 	
 	// local 
 	float peakHold[2] = {0.0f, 0.0f};
-	float holdTimeRemainBeforeReset = 0.0f;
+	double holdTimeRemainBeforeReset = 0.0;
 	float yellowThreshold;// in px, before vertical inversion
 	float redThreshold;// in px, before vertical inversion
 	int colorTheme;
